removed ID
found your link to the ebuild in https://bugs.gentoo.org/525882#c50
Are you interested in preparing a pull request with the help of the proxy maintainers? Join us on IRC, if you like. 
Some ideas for a start:
What needs to be done with the ebuild: You should test this ebuild with repoman, bump to EAPI=6
change epatch to PATCH= or eapply see https://blogs.gentoo.org/mgorny/2015/11/13/the-ultimate-guide-to-eapi-6/ and https://devmanual.gentoo.org/ebuild-writing/eapi/
There are some absolute paths in the ebuild which need to be checked
sz in line 104 needs local before